Configure and Reorganize Documentation Generation
- Use configure_file() to configure the project name, version, and
paths in the doxygen input files.
- Add a make target for docs, allowing create_docs.sh to be deleted.
- Move doxygen input files under the src directory and add
CMakeLists.txt to configure them,
- Remove the contents of the original doc directory.
- Modify the README file with new instructions for generating docs.
- Add a README_FIRST to src/doc to cover the possibility that doxywizard
or some other tool may strip the license information from the files.
- Split the generated documentation to reflect the different client and
module APIs.
- In keeping with supporting out-of-tree builds, documentation is generated
under doc/ in the build directory. But the user can override this on the
cmake command line.
- Updated the .gitignore file with a license block, and modified the list of
files - including other variants for the build directory name and the
original doc/ directory in the project root.
- Note the prepending of a "/" to directory names in .gitignore. This
stops the directories being ignored in subdirectories. Ignoring /doc
in the root also ignored /src/doc!