Eliminate possible buffer overrun
:Release Notes:
Corrected a potential buffer overflow detected by coverity.

:Detailed Notes:
nyx_utils_write_value was using a buffer which was too small to hold the
maximum length string plus the terminating nul, leading to a potential
buffer overflow.

The buffer size was doubled to ensure it could hold any possible value,
snprintf is used instead of sprintf, the provided length always leaves
room for the addition of a terminating nul, and a nul is added to ensure
that a truncated conversion is always terminated.

A similar set of changes were made to nyx_utils_read_value, to ensure
that a truncated read is always nul-terminated.

:Testing Performed:
Clean compile and verification by inspection

:QA Notes:

:Issues Addressed:
[GF-37400] nyx-lib: Coverity Issue

Open-webOS-DCO-1.0-Signed-off-by: Keith Derrick <keith.derrick@lge.com>

Change-Id: I62485d0bc017fefc8e17e1fdc605411641ac41da
Reviewed-on: https://g2g.palm.com/3396
Reviewed-by: Build Verification
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Tested-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Ameya Palande <ameya.palande@lge.com>
Reviewed-by: Sapna Todwal <sapna.todwal@lge.com>